URI Parser 2.0.0: made field decoding/encoding separate
This way the HTML filter will leave URIs as they are (otherwise some
of them break because of very poor handling of encoding by other
languages / services), unless the user wants to modify them.

Also changed the URL text encoding to not encode "/" as that is not
handled correctly by some services as well.